Add spending management feature in Myfinance app
This commit introduces a bunch of new classes, interfaces and properties related to the spending management feature in the Myfinance application.  This includes creating new Entity classes (Rule, Spending, Transfer), their corresponding DTOs and Repositories, as well as the necessary Services and Tests. The application setup and properties are also defined.