Refactor FinanceOverviewController and apply UI changes
The FinanceOverviewController file has been reframed for better maintainability through restructuring and removal of redundant code. Additional services and converters were implemented to help with this. New features were also introduced to the UI, including the addition of overview and new spending pages, and the integration of new CSS styles. Dependency on ThymeLeaf added to pom.xml. Several other code and test files updated in the process.