Add new table headers and modify table data in templates
Added "Action" header in the overview.html and rules.html templates, as well as "Status" header in rules.html and transfer.html templates. Modified table data in the rules.html and transfer.html templates to improve clarity. Updated Transformer and SpendingOverviewDto.java to replace TransferDto with a simple transferDescription string.