Update rule and transfer management in Spending services
Implemented changes in the Spending services to manage rules and transfers. Now, the SpendingEditService, SpendingDeletionService, and SpendingCreatorService properly handle the addition, removal, and update of rules and transfers associated with a Spending. Also, introduced new services RuleUpdateService and TransferUpdateService to handle rule and transfer updates directly. Removed unnecessary transactional annotations on Entities.