Refactor code for transfer changing process and remove redundant methods
The commit has reorganized code in SpendingEditService to make sure that the old transfer is obtained before trying to replace it with a new one. A new method has been added in TransferChangeService to check if a transfer's amount has been modified. Additionally, unnecessary methods in SpendingRepository that were retrieving all spendings with a certain transfer ID or with an amount greater than specified, have been removed. Some small test annotation rearrangement was also done in SpendingDeletionServiceTest.